Move WooAnalyticsStat to WooFoundation to reuse between POS and WooApp
