+++ Modules/Sources/WooFoundationCore/Analytics/WooAnalyticsEventPropertyType.swift
