+++ Modules/Sources/WooFoundationCore/Analytics/WooAnalyticsStat.swift
@@ -1,6 +1,4 @@
-#if canImport(WordPressShared)
-import WordPressShared
-#endif
+import Foundation
 
 /// This enum contains all of the events we track in the WooCommerce app.
 ///
@@ -18,7 +16,7 @@ import WordPressShared
 /// Note: If you would like to exclude site properties (e.g. `blog_id`) for a given event, please
 /// add the event to the `WooAnalyticsStat.shouldSendSiteProperties` var.
 ///
-enum WooAnalyticsStat: String {
+public enum WooAnalyticsStat: String {
 
     // MARK: Application Events
     //
@@ -1357,7 +1355,7 @@ extension WooAnalyticsStat {
     /// Note: Currently all authentication events will return false. If you wish
     /// to include additional no-site-info events, please add them here.
     ///
-    var shouldSendSiteProperties: Bool {
+    public var shouldSendSiteProperties: Bool {
         switch self {
         // Authentication Events
         case .signedIn, .logout, .openedLogin, .loginFailed,
@@ -1379,102 +1377,3 @@ extension WooAnalyticsStat {
         }
     }
 }
-
-#if canImport(WordPressShared)
-extension WooAnalyticsStat {
-
-    /// Converts the provided WPAnalyticsStat into a WooAnalyticsStat.
-    /// This whole process kinda stinks, but we need this for the `WordPressAuthenticatorDelegate`
-    /// implementation. ☹️ Feel free to refactor later on!
-    ///
-    /// - Parameter stat: The WPAnalyticsStat to convert
-    /// - Returns: The corresponding WooAnalyticsStat or nil if it cannot be converted
-    ///
-    static func valueOf(stat: WPAnalyticsStat) -> WooAnalyticsStat? {
-        var wooEvent: WooAnalyticsStat? = nil
-
-        switch stat {
-        case .signedIn:
-            wooEvent = WooAnalyticsStat.signedIn
-        case .signedInToJetpack:
-            wooEvent = WooAnalyticsStat.signedIn
-        case .logout:
-            wooEvent = WooAnalyticsStat.logout
-        case .openedLogin:
-            wooEvent = WooAnalyticsStat.openedLogin
-        case .loginFailed:
-            wooEvent = WooAnalyticsStat.loginFailed
-        case .loginAutoFillCredentialsFilled:
-            wooEvent = WooAnalyticsStat.loginAutoFillCredentialsFilled
-        case .loginAutoFillCredentialsUpdated:
-            wooEvent = WooAnalyticsStat.loginAutoFillCredentialsUpdated
-        case .loginProloguePaged:
-            wooEvent = WooAnalyticsStat.loginProloguePaged
-        case .loginPrologueViewed:
-            wooEvent = WooAnalyticsStat.loginPrologueViewed
-        case .loginEmailFormViewed:
-            wooEvent = WooAnalyticsStat.loginEmailFormViewed
-        case .loginMagicLinkOpenEmailClientViewed:
-            wooEvent = WooAnalyticsStat.loginMagicLinkOpenEmailClientViewed
-        case .loginMagicLinkRequestFormViewed:
-            wooEvent = WooAnalyticsStat.loginMagicLinkRequestFormViewed
-        case .loginMagicLinkExited:
-            wooEvent = WooAnalyticsStat.loginMagicLinkExited
-        case .loginMagicLinkFailed:
-            wooEvent = WooAnalyticsStat.loginMagicLinkFailed
-        case .loginMagicLinkOpened:
-            wooEvent = WooAnalyticsStat.loginMagicLinkOpened
-        case .loginMagicLinkRequested:
-            wooEvent = WooAnalyticsStat.loginMagicLinkRequested
-        case .loginMagicLinkSucceeded:
-            wooEvent = WooAnalyticsStat.loginMagicLinkSucceeded
-        case .loginPasswordFormViewed:
-             wooEvent = WooAnalyticsStat.loginPasswordFormViewed
-        case .loginURLFormViewed:
-            wooEvent = WooAnalyticsStat.loginURLFormViewed
-        case .loginURLHelpScreenViewed:
-            wooEvent = WooAnalyticsStat.loginURLHelpScreenViewed
-        case .loginUsernamePasswordFormViewed:
-            wooEvent = WooAnalyticsStat.loginUsernamePasswordFormViewed
-        case .loginTwoFactorFormViewed:
-            wooEvent = WooAnalyticsStat.loginTwoFactorFormViewed
-        case .loginEpilogueViewed:
-            wooEvent = WooAnalyticsStat.loginEpilogueViewed
-        case .loginForgotPasswordClicked:
-            wooEvent = WooAnalyticsStat.loginForgotPasswordClicked
-        case .loginSocialButtonClick:
-            wooEvent = WooAnalyticsStat.loginSocialButtonClick
-        case .loginSocialButtonFailure:
-            wooEvent = WooAnalyticsStat.loginSocialButtonFailure
-        case .loginSocialConnectSuccess:
-            wooEvent = WooAnalyticsStat.loginSocialConnectSuccess
-        case .loginSocialConnectFailure:
-            wooEvent = WooAnalyticsStat.loginSocialConnectFailure
-        case .loginSocialSuccess:
-            wooEvent = WooAnalyticsStat.loginSocialSuccess
-        case .loginSocialFailure:
-            wooEvent = WooAnalyticsStat.loginSocialFailure
-        case .loginSocial2faNeeded:
-            wooEvent = WooAnalyticsStat.loginSocial2faNeeded
-        case .loginSocialAccountsNeedConnecting:
-            wooEvent = WooAnalyticsStat.loginSocialAccountsNeedConnecting
-        case .loginSocialErrorUnknownUser:
-            wooEvent = WooAnalyticsStat.loginSocialErrorUnknownUser
-        case .onePasswordFailed:
-            wooEvent = WooAnalyticsStat.onePasswordFailed
-        case .onePasswordLogin:
-            wooEvent = WooAnalyticsStat.onePasswordLogin
-        case .onePasswordSignup:
-            wooEvent = WooAnalyticsStat.onePasswordSignup
-        case .twoFactorCodeRequested:
-            wooEvent = WooAnalyticsStat.twoFactorCodeRequested
-        case .twoFactorSentSMS:
-            wooEvent = WooAnalyticsStat.twoFactorSentSMS
-        default:
-            wooEvent = nil
-        }
-
-        return wooEvent
-    }
-}
-#endif
