+++ WooCommerce/Classes/Analytics/WooAnalyticsEvent+WooApp.swift
@@ -3,54 +3,8 @@ import UIKit
 import Yosemite
 import WooFoundation
 
-/// This struct represents an analytics event. It is a combination of `WooAnalyticsStat` and
-/// its properties.
-///
-/// This was mostly created to promote static-typing via constructors.
-///
-/// ## Adding New Events
-///
-/// 1. Add the event name (`String`) to `WooAnalyticsStat`.
-/// 2. Create an `extension` of `WooAnalyticsStat` if necessary for grouping.
-/// 3. Add a `static func` constructor.
-///
-/// Here is an example:
-///
-/// ~~~
-/// extension WooAnalyticsEvent {
-///     enum LoginStep: String {
-///         case start
-///         case success
-///     }
-///
-///     static func login(step: LoginStep) -> WooAnalyticsEvent {
-///         let properties = [
-///             "step": step.rawValue
-///         ]
-///
-///         return WooAnalyticsEvent(name: "login", properties: properties)
-///     }
-/// }
-/// ~~~
-///
-/// Examples of tracking calls (in the client App or Pod):
-///
-/// ~~~
-/// Analytics.track(event: .login(step: .start))
-/// Analytics.track(event: .loginStart)
-/// ~~~
-///
-struct WooAnalyticsEvent {
-    init(statName: WooAnalyticsStat, properties: [String: WooAnalyticsEventPropertyType], error: Error? = nil) {
-        self.statName = statName
-        self.properties = properties
-        self.error = error
-    }
-
-    let statName: WooAnalyticsStat
-    let properties: [String: WooAnalyticsEventPropertyType]
-    let error: Error?
-}
+typealias WooAnalyticsEvent = WooFoundation.WooAnalyticsEvent
+typealias WooAnalyticsStat = WooFoundation.WooAnalyticsStat
 
 // MARK: - In-app Feedback and Survey
